cachalka/requirements.txt
Django==1.8.4
django-model-utils==2.3.1